backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
opencv-python
numpy
torch
torchvision
pillow
python-multipart